fix: allow browser helpers to be loaded from a browser build
